Broke out components into reusable py modules
