+++ src/aac/ui/components/BaseTkinterComponent.py
@@ -1,0 +1,11 @@
+from attr import attrs, attrib
+
+
+@attrs
+class BaseTkinterComponent:
+    """Base class for all custom Tkinter components."""
+
+    parent_widget = attrib()
+
+    def build(self):
+        raise NotImplementedError()
